pyproject.toml
[project]
name = "pdfserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httptools>=0.7.1",
    "hypercorn>=0.18.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyhanko>=0.32.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.7",
    "watchgod>=0.8.2",
]
